server/pyproject.toml
[project]
name = "bayesmech_camalytics"
version = "0.1.0"
description = "BayesMech CamAlytics - AR Stream Server with SAM3 Video Segmentation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "websockets==12.0",
    # Serialization
    "protobuf==4.25.1",
    # Image processing
    "Pillow==10.2.0",
    "opencv-python>=4.9.0",
    # Array operations and ML
    "numpy==1.26.3",
    "torch==2.7.0",
    "torchvision>=0.22.0",
    "torchaudio>=2.7.0",
    # SAM2/SAM3 and dependencies
    "SAM-2 @ git+https://github.com/facebookresearch/sam2.git",
    "matplotlib>=3.8.0",
    "scikit-learn>=1.4.0",
    # Configuration
    "pyyaml==6.0.1",
    # Utilities
    "python-multipart==0.0.6",
    "aiohttp>=3.13.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["server", "segmentation"]

[tool.uv]
index-url = "https://download.pytorch.org/whl/cu126"
extra-index-url = ["https://pypi.org/simple"]
